refactor: combine Copy To and Copy From
