+++ src/datanode/src/sql/copy_table_to.rs
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::HashMap;
 use std::pin::Pin;
 
+use common_datasource;
+use common_datasource::object_store::{build_backend, parse_url};
 use common_query::physical_plan::SessionContext;
 use common_query::Output;
 use common_recordbatch::adapter::DfRecordBatchStreamAdapter;
@@ -27,51 +28,12 @@ use object_store::ObjectStore;
 use snafu::ResultExt;
 use table::engine::TableReference;
 use table::requests::CopyTableRequest;
-use url::{ParseError, Url};
 
-use super::copy_table_from::{build_fs_backend, build_s3_backend, S3_SCHEMA};
 use crate::error::{self, Result};
 use crate::sql::SqlHandler;
 
 impl SqlHandler {
-    fn build_backend(
-        &self,
-        url: &str,
-        connection: HashMap<String, String>,
-    ) -> Result<(ObjectStore, String)> {
-        let result = Url::parse(url);
-
-        match result {
-            Ok(url) => {
-                let host = url.host_str();
-
-                let schema = url.scheme();
-
-                let path = url.path();
-
-                match schema.to_uppercase().as_str() {
-                    S3_SCHEMA => {
-                        let object_store = build_s3_backend(host, "/", connection)?;
-                        Ok((object_store, path.to_string()))
-                    }
-
-                    _ => error::UnsupportedBackendProtocolSnafu {
-                        protocol: schema.to_string(),
-                    }
-                    .fail(),
-                }
-            }
-            Err(ParseError::RelativeUrlWithoutBase) => {
-                let object_store = build_fs_backend("/")?;
-                Ok((object_store, url.to_string()))
-            }
-            Err(err) => Err(error::Error::InvalidUrl {
-                url: url.to_string(),
-                source: err,
-            }),
-        }
-    }
-    pub(crate) async fn copy_table(&self, req: CopyTableRequest) -> Result<Output> {
+    pub(crate) async fn copy_table_to(&self, req: CopyTableRequest) -> Result<Output> {
         let table_ref = TableReference {
             catalog: &req.catalog_name,
             schema: &req.schema_name,
@@ -91,9 +53,11 @@ impl SqlHandler {
             .context(error::TableScanExecSnafu)?;
         let stream = Box::pin(DfRecordBatchStreamAdapter::new(stream));
 
-        let (object_store, file_name) = self.build_backend(&req.file_name, req.connection)?;
+        let (_schema, _host, path) = parse_url(&req.location).context(error::ParseUrlSnafu)?;
+        let object_store =
+            build_backend(&req.location, req.connection).context(error::BuildBackendSnafu)?;
 
-        let mut parquet_writer = ParquetWriter::new(file_name, stream, object_store);
+        let mut parquet_writer = ParquetWriter::new(path.to_string(), stream, object_store);
         // TODO(jiachun):
         // For now, COPY is implemented synchronously.
         // When copying large table, it will be blocked for a long time.
